Move tasks to import export unit test
